// frontend/src/types/proto-es/v1/idp_service_pb.d.ts
export declare type IdentityProvider = Message<"laelia.v1.IdentityProvider"> & {
  /**
   * The name of the identity provider.
   * Format: idps/{idp}
   *
   * @generated from field: string name = 1;
   */
  name: string;

  /**
   * The display title of the identity provider.
   *
   * @generated from field: string title = 2;
   */
  title: string;

  /**
   * The domain for email matching when using this identity provider.
   *
   * @generated from field: string domain = 3;
   */
  domain: string;

  /**
   * The type of identity provider protocol.
   *
   * @generated from field: laelia.v1.IdentityProviderType type = 4;
   */
  type: IdentityProviderType;

  /**
   * The configuration details for the identity provider.
   *
   * @generated from field: laelia.v1.IdentityProviderConfig config = 5;
   */
  config?: IdentityProviderConfig | undefined;
};

/**
 * Describes the message laelia.v1.IdentityProvider.
 * Use `create(IdentityProviderSchema)` to create a new message.
 */
export declare const IdentityProviderSchema: GenMessage<IdentityProvider>;

/**
 * @generated from message laelia.v1.IdentityProviderConfig
 */
export declare type IdentityProviderConfig = Message<"laelia.v1.IdentityProviderConfig"> & {
  /**
   * @generated from oneof laelia.v1.IdentityProviderConfig.config
   */
  config: {
    /**
     * OAuth2 protocol configuration.
     *
     * @generated from field: laelia.v1.OAuth2IdentityProviderConfig oauth2_config = 1;
     */
    value: OAuth2IdentityProviderConfig;
    case: "oauth2Config";
  } | {
    /**
     * OIDC protocol configuration.
     *
     * @generated from field: laelia.v1.OIDCIdentityProviderConfig oidc_config = 2;
     */
    value: OIDCIdentityProviderConfig;
    case: "oidcConfig";
  } | {
    /**
     * LDAP protocol configuration.
     *
     * @generated from field: laelia.v1.LDAPIdentityProviderConfig ldap_config = 3;
     */
    value: LDAPIdentityProviderConfig;
    case: "ldapConfig";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message laelia.v1.IdentityProviderConfig.
 * Use `create(IdentityProviderConfigSchema)` to create a new message.
 */
export declare const IdentityProviderConfigSchema: GenMessage<IdentityProviderConfig>;

/**
 * OAuth2IdentityProviderConfig is the structure for OAuth2 identity provider config.
 *
 * @generated from message laelia.v1.OAuth2IdentityProviderConfig
 */
export declare type OAuth2IdentityProviderConfig = Message<"laelia.v1.OAuth2IdentityProviderConfig"> & {
  /**
   * The authorization endpoint URL for OAuth2 flow.
   *
   * @generated from field: string auth_url = 1;
   */
  authUrl: string;

  /**
   * The token endpoint URL for exchanging authorization code.
   *
   * @generated from field: string token_url = 2;
   */
  tokenUrl: string;

  /**
   * The user information endpoint URL.
   *
   * @generated from field: string user_info_url = 3;
   */
  userInfoUrl: string;

  /**
   * The OAuth2 client identifier.
   *
   * @generated from field: string client_id = 4;
   */
  clientId: string;

  /**
   * The OAuth2 client secret for authentication.
   *
   * @generated from field: string client_secret = 5;
   */
  clientSecret: string;

  /**
   * The list of OAuth2 scopes to request.
   *
   * @generated from field: repeated string scopes = 6;
   */
  scopes: string[];

  /**
   * Mapping configuration for user attributes from OAuth2 response.
   *
   * @generated from field: laelia.v1.FieldMapping field_mapping = 7;
   */
  fieldMapping?: FieldMapping | undefined;

  /**
   * Whether to skip TLS certificate verification.
   *
   * @generated from field: bool skip_tls_verify = 8;
   */
  skipTlsVerify: boolean;

  /**
   * The authentication style for client credentials.
   *
   * @generated from field: laelia.v1.OAuth2AuthStyle auth_style = 9;
   */
  authStyle: OAuth2AuthStyle;
};
// ...
export declare type OIDCIdentityProviderConfig = Message<"laelia.v1.OIDCIdentityProviderConfig"> & {
  /**
   * The OIDC issuer URL for the identity provider.
   *
   * @generated from field: string issuer = 1;
   */
  issuer: string;

  /**
   * The OIDC client identifier.
   *
   * @generated from field: string client_id = 2;
   */
  clientId: string;

  /**
   * The OIDC client secret for authentication.
   *
   * @generated from field: string client_secret = 3;
   */
  clientSecret: string;

  /**
   * The scopes that the OIDC provider supports.
   *
   * @generated from field: repeated string scopes = 4;
   */
  scopes: string[];

  /**
   * Mapping configuration for user attributes from OIDC claims.
   *
   * @generated from field: laelia.v1.FieldMapping field_mapping = 5;
   */
  fieldMapping?: FieldMapping | undefined;

  /**
   * Whether to skip TLS certificate verification.
   *
   * @generated from field: bool skip_tls_verify = 6;
   */
  skipTlsVerify: boolean;

  /**
   * The authentication style for client credentials.
   *
   * @generated from field: laelia.v1.OAuth2AuthStyle auth_style = 7;
   */
  authStyle: OAuth2AuthStyle;
};
// ...
export declare type LDAPIdentityProviderConfig = Message<"laelia.v1.LDAPIdentityProviderConfig"> & {
  /**
   * Host is the hostname or IP address of the LDAP server, e.g.,
   * "ldap.example.com".
   *
   * @generated from field: string host = 1;
   */
  host: string;

  /**
   * Port is the port number of the LDAP server, e.g., 389. When not set, the
   * default port of the corresponding security protocol will be used, i.e. 389
   * for StartTLS and 636 for LDAPS.
   *
   * @generated from field: int32 port = 2;
   */
  port: number;

  /**
   * SkipTLSVerify controls whether to skip TLS certificate verification.
   *
   * @generated from field: bool skip_tls_verify = 3;
   */
  skipTlsVerify: boolean;

  /**
   * BindDN is the DN of the user to bind as a service account to perform
   * search requests.
   *
   * @generated from field: string bind_dn = 4;
   */
  bindDn: string;

  /**
   * BindPassword is the password of the user to bind as a service account.
   *
   * @generated from field: string bind_password = 5;
   */
  bindPassword: string;

  /**
   * BaseDN is the base DN to search for users, e.g., "ou=users,dc=example,dc=com".
   *
   * @generated from field: string base_dn = 6;
   */
  baseDn: string;

  /**
   * UserFilter is the filter to search for users, e.g., "(uid=%s)".
   *
   * @generated from field: string user_filter = 7;
   */
  userFilter: string;

  /**
   * SecurityProtocol is the security protocol to be used for establishing
   * connections with the LDAP server.
   *
   * @generated from field: laelia.v1.LDAPIdentityProviderConfig.SecurityProtocol security_protocol = 8;
   */
  securityProtocol: LDAPIdentityProviderConfig_SecurityProtocol;

  /**
   * FieldMapping is the mapping of the user attributes returned by the LDAP
   * server.
   *
   * @generated from field: laelia.v1.FieldMapping field_mapping = 9;
   */
  fieldMapping?: FieldMapping | undefined;
};
// ...
export enum IdentityProviderType {
  /**
   * Unspecified identity provider type.
   *
   * @generated from enum value: IDENTITY_PROVIDER_TYPE_UNSPECIFIED = 0;
   */
  IDENTITY_PROVIDER_TYPE_UNSPECIFIED = 0,

  /**
   * OAuth 2.0 authentication protocol.
   *
   * @generated from enum value: OAUTH2 = 1;
   */
  OAUTH2 = 1,

  /**
   * OpenID Connect authentication protocol.
   *
   * @generated from enum value: OIDC = 2;
   */
  OIDC = 2,

  /**
   * LDAP directory service authentication.
   *
   * @generated from enum value: LDAP = 3;
   */
  LDAP = 3,
}